Unit testing for GitHub APIs
